+++ src/utils/functions/dfs.h
@@ -1,11 +1,11 @@
 #ifndef DFS_H
 #define DFS_H
 
-void DFS(int** graph, int &n, int &vertex, bool* visited, int& componentSize, int& componentEdges) {
+void DFS(lli** graph, lli &n, lli &vertex, bool* visited, lli& componentSize, lli& componentEdges) {
     visited[vertex] = true;
     ++componentSize;
 
-    for (int i = 0; i < n; ++i) {
+    for (lli i = 0; i < n; ++i) {
         if (graph[vertex][i] && !visited[i]) {
             DFS(graph, n, i, visited, componentSize, componentEdges);
             componentEdges += graph[vertex][i];
